refactor: button -> link 로 수정
